+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - 1.1.3
   - Fixed console warnings
   - Fixed dev hot reload: now state is preserving
+  - Added render hoooks to `config.settings`: `renderBeforeWidget`, `renderAfterWidget`, `renderBeforeActions`, `renderAfterActions`
 - 1.1.2
   - Added import from [JsonLogic](http://jsonlogic.com)
 - 1.1.1